Extract tracking and cleanup logic out to separate file
